# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TechZone Electronics & Gaming Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/track:
    get:
      operationId: trackOrder
      tags: [orders]
      summary: Track order by ID and phone
      parameters:
        - name: orderId
          in: query
          required: true
          schema: { type: string }
        - name: phone
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderTracking"
        "404":
          description: Not found
  /newsletter:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "200":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterResponse"
...
  /customers/subscriptions:
    get:
      operationId: customerSubscriptions
      tags: [customers]
      summary: Current customer subscriptions
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerSubscription"
    post:
      operationId: createCustomerSubscription
      tags: [customers]
      summary: Subscribe to a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSubscription"
components:
  schemas:
...
    OrderItem:
      type: object
      required: [productId, nameAr, price, quantity, image]
      properties:
        productId: { type: string }
        nameAr:    { type: string }
        price:     { type: number }
        quantity:  { type: integer }
        image:     { type: string }
    OrderInput:
      type: object
      required: [customerName, phone, city, address, items, paymentMethod]
      properties:
        customerName:  { type: string }
        phone:         { type: string }
        city:          { type: string }
        address:       { type: string }
        email:         { type: ["string", "null"] }
        promoCode:     { type: ["string", "null"] }
        notes:         { type: ["string", "null"] }
        paymentMethod: { type: string, enum: [cash_on_delivery, card, bank_transfer] }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    Order:
      type: object
      required: [id, customerName, phone, city, address, items, paymentMethod, subtotal, shipping, total, status, createdAt]
      properties:
        id:            { type: string }
        customerName:  { type: string }
        phone:         { type: string }
        city:          { type: string }
        address:       { type: string }
        email:         { type: ["string", "null"] }
        promoCode:     { type: ["string", "null"] }
        discount:      { type: ["number", "null"] }
        notes:         { type: ["string", "null"] }
        paymentMethod: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:  { type: number }
        shipping:  { type: number }
        total:     { type: number }
        status:    { type: string }
        createdAt: { type: string }
    OrderTracking:
      type: object
      required: [id, status, timeline, estimatedDelivery]
      properties:
        id:     { type: string }
        status: { type: string }
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TrackingStep"
        estimatedDelivery: { type: string }
        trackingNumber:    { type: ["string", "null"] }
    TrackingStep:
      type: object
      required: [statusAr, completed, date]
      properties:
        statusAr:  { type: string }
        completed: { type: boolean }
        date:      { type: ["string", "null"] }
    NewsletterInput:
      type: object
      required: [email]
      properties:
        email: { type: string }
    NewsletterResponse:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string }
...
    CustomerSubscription:
      type: object
      required: [id, customerId, planId, status, renewCount, createdAt]
      properties:
        id:          { type: string }
        customerId:  { type: string }
        planId:      { type: string }
        status:      { type: string, enum: [active, cancelled, expired, pending] }
        startedAt:   { type: ["string", "null"] }
        expiresAt:   { type: ["string", "null"] }
        renewCount:  { type: integer }
        cancelledAt: { type: ["string", "null"] }
        createdAt:   { type: string }
    CreateSubscriptionInput:
      type: object
      required: [planId]
      properties:
        planId: { type: string }